deskulpt_plugin

Attribute Macro dispatch

#[dispatch]

Dispatch a Deskulpt plugin command.

The PluginCommand::run method requires the serde_json::Value type for command input and output so as to interoperate with calls from the widgets in the frontend. This would require manual deserialization and serialization when implementing any command.

When marked with #[dispatch], the signature of the method remains the same, except that input is allowed to be any type that implements [serde::Deserialize], and the return type is allowed to be Result<T, E> for any type T that implements [serde::Serialize] and any type E that can be converted to anyhow::Error with the ? operator. That said, the most convenient way would be to use anyhow::Result<T> for the return type directly.

Example

use anyhow::Result;
use deskulpt_plugin::{dispatch, EngineInterface, PluginCommand};
use serde::{Deserialize, Serialize};

// Implement the plugin...

struct MetadataCommand;

#[derive(Deserialize)]
struct InputPayload {
    path: std::path::PathBuf,
}

#[derive(Serialize)]
struct OutputPayload {
    is_dir: bool,
    is_file: bool,
    is_symlink: bool,
    len: u64,
}

impl PluginCommand for MetadataCommand {
    // Associate types and methods...

    #[dispatch]
    fn run(
        &self,
        _id: String,
        _plugin: &Self::Plugin,
        _engine: &EngineInterface,
        input: InputPayload,     // Custom deserializable input type
    ) -> Result<OutputPayload> { // Custom serializable output type
        let metadata = std::fs::metadata(input.path)?;
        Ok(OutputPayload {
            is_dir: metadata.is_dir(),
            is_file: metadata.is_file(),
            is_symlink: metadata.file_type().is_symlink(),
            len: metadata.len(),
        })
    }
}