deskulpt_plugin

Function call_plugin

pub fn call_plugin<P: Plugin>(
    widget_dir_fn: impl Fn(&str) -> Result<PathBuf> + 'static,
    plugin: &P,
    command: &str,
    id: String,
    payload: Option<Value>,
) -> Result<Value>

Call a Deskulpt plugin (🚧 TODO 🚧).

🚧 TODO 🚧

This function should be completed removed and replaced with a serve_plugin function that will serve as the entry point of the plugin, running it as a standalone process that can interact with the Deskulpt core through IPC. See nushell for reference.